DRAW-172-1 refactor: 코드 리뷰 반영
