+++ core/src/main/kotlin/com/xorker/draw/mafia/MafiaGameRoomService.kt
@@ -3,9 +3,9 @@ package com.xorker.draw.mafia
 import com.xorker.draw.room.Room
 import com.xorker.draw.websocket.Session
 import com.xorker.draw.websocket.SessionEventListener
-import org.springframework.stereotype.Component
+import org.springframework.stereotype.Service
 
-@Component
+@Service
 internal class MafiaGameRoomService(
     private val mafiaGameRepository: MafiaGameRepository,
     private val mafiaGameMessenger: MafiaGameMessenger,
