+++ core/src/main/kotlin/com/xorker/draw/mafia/MafiaGameService.kt
@@ -3,17 +3,20 @@ package com.xorker.draw.mafia
 import com.xorker.draw.exception.InvalidRequestOnlyMyTurnException
 import com.xorker.draw.exception.InvalidRequestValueException
 import com.xorker.draw.mafia.dto.DrawRequest
-import com.xorker.draw.timer.TimerRepository
+import com.xorker.draw.mafia.phase.MafiaPhasePlayGameProcessor
+import com.xorker.draw.mafia.phase.MafiaPhaseService
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.Session
-import org.springframework.stereotype.Component
+import java.util.*
+import org.springframework.stereotype.Service
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.contract
 
-@Component
+@Service
 internal class MafiaGameService(
+    private val mafiaPhaseService: MafiaPhaseService,
+    private val mafiaPhasePlayGameProcessor: MafiaPhasePlayGameProcessor,
     private val mafiaGameRepository: MafiaGameRepository,
-    private val timerRepository: TimerRepository,
     private val mafiaGameMessenger: MafiaGameMessenger,
 ) : MafiaGameUseCase {
 
@@ -32,51 +35,45 @@ internal class MafiaGameService(
         mafiaGameMessenger.broadcastDraw(gameInfo.room.id, request.drawData)
     }
 
-    override fun nextTurnByUser(session: Session, nextStep: () -> Unit) {
+    override fun nextTurnByUser(session: Session) {
         val gameInfo = session.getGameInfo()
         val phase = gameInfo.phase
         assertTurn(phase, session.user.id)
 
         phase.timerJob.cancel()
 
-        processNextTurn(gameInfo, nextStep)
+        mafiaPhasePlayGameProcessor.processNextTurn(gameInfo) {
+            mafiaPhaseService.vote(gameInfo.room.id)
+        }
     }
 
-    internal fun playMafiaGame(gameInfo: MafiaGameInfo, nextStep: () -> Unit): MafiaPhase.Playing {
-        val phase = gameInfo.phase
-        assertIs<MafiaPhase.Ready>(phase)
+    override fun voteMafia(session: Session, targetUserId: UserId) {
+        val gameInfo = session.getGameInfo()
 
-        val gameOption = gameInfo.gameOption
+        val voter = session.user
+        val voterUserId = voter.id
 
-        val job = timerRepository.startTimer(gameOption.turnTime) {
-            processNextTurn(gameInfo, nextStep)
-        }
+        assertIs<MafiaPhase.Vote>(gameInfo.phase)
+        val phase = gameInfo.phase as MafiaPhase.Vote
 
-        val playingPhase = phase.toPlaying(job)
-        gameInfo.phase = playingPhase
+        vote(phase.players, voterUserId, targetUserId)
 
-        return playingPhase
+        mafiaGameMessenger.broadcastVoteStatus(gameInfo)
     }
 
-    private fun processNextTurn(gameInfo: MafiaGameInfo, nextStep: () -> Unit) {
-        val phase = gameInfo.phase
-        assertIs<MafiaPhase.Playing>(phase)
+    private fun vote(
+        players: Map<UserId, Vector<UserId>>,
+        voterUserId: UserId,
+        targetUserId: UserId,
+    ) {
+        players.forEach { player ->
+            val userIds = player.value
 
-        val gameOption = gameInfo.gameOption
-
-        val nextTurn = phase.nextTurn(gameOption.round, gameOption.turnCount)
-
-        if (nextTurn == null) {
-            nextStep.invoke()
-            return
+            if (voterUserId in userIds) {
+                userIds.remove(voterUserId)
+            }
         }
-
-        phase.turnInfo = nextTurn
-
-        mafiaGameMessenger.broadcastNextTurn(gameInfo)
-        phase.timerJob = timerRepository.startTimer(gameInfo.gameOption.turnTime) {
-            processNextTurn(gameInfo, nextStep)
-        }
+        players[targetUserId]?.add(voterUserId) ?: InvalidRequestValueException
     }
 
     private fun Session.getGameInfo(): MafiaGameInfo =
