+++ core/src/main/kotlin/com/xorker/draw/mafia/MafiaGameUseCase.kt
@@ -1,9 +1,11 @@
 package com.xorker.draw.mafia
 
 import com.xorker.draw.mafia.dto.DrawRequest
+import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.Session
 
 interface MafiaGameUseCase {
     fun draw(session: Session, request: DrawRequest)
-    fun nextTurnByUser(session: Session, nextStep: () -> Unit)
+    fun nextTurnByUser(session: Session)
+    fun voteMafia(session: Session, targetUserId: UserId)
 }
