--- core/src/main/kotlin/com/xorker/draw/mafia/MafiaVoteUseCase.kt
@@ -1,8 +1,0 @@
-package com.xorker.draw.mafia
-
-import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.Session
-
-interface MafiaVoteUseCase {
-    fun voteMafia(session: Session, targetUserId: UserId)
-}
